[fixed] loot/threat map [upadted] for new serialization needs db wipe
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@4402 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/managers/ZoneManager.cpp
@@ -134,7 +134,7 @@ void ZoneManagerImplementation::readObject(ObjectInputStream* stream) {
 		String _name;
 		_name.parseFromBinaryStream(stream);
 
-		uint16 _varSize = stream->readShort();
+		uint32 _varSize = stream->readInt();
 
 		int _currentOffset = stream->getOffset();
 
@@ -167,7 +167,7 @@ int ZoneManagerImplementation::writeObjectMembers(ObjectOutputStream* stream) {
 
 	String _name;
 	int _offset;
-	uint16 _totalSize;
+	uint32 _totalSize;
 
 	return _count + 0;
 }
