+++ MMOCoreORB/src/server/zone/managers/loot/LootManager.cpp
@@ -209,7 +209,7 @@ void LootManagerImplementation::readObject(ObjectInputStream* stream) {
 		String _name;
 		_name.parseFromBinaryStream(stream);
 
-		uint16 _varSize = stream->readShort();
+		uint32 _varSize = stream->readInt();
 
 		int _currentOffset = stream->getOffset();
 
@@ -242,7 +242,7 @@ int LootManagerImplementation::writeObjectMembers(ObjectOutputStream* stream) {
 
 	String _name;
 	int _offset;
-	uint16 _totalSize;
+	uint32 _totalSize;
 
 	return _count + 0;
 }
