+++ MMOCoreORB/src/server/zone/objects/installation/harvester/HarvesterObject.cpp
@@ -205,7 +205,7 @@ void HarvesterObjectImplementation::readObject(ObjectInputStream* stream) {
 		String _name;
 		_name.parseFromBinaryStream(stream);
 
-		uint16 _varSize = stream->readShort();
+		uint32 _varSize = stream->readInt();
 
 		int _currentOffset = stream->getOffset();
 
@@ -238,7 +238,7 @@ int HarvesterObjectImplementation::writeObjectMembers(ObjectOutputStream* stream
 
 	String _name;
 	int _offset;
-	uint16 _totalSize;
+	uint32 _totalSize;
 
 	return _count + 0;
 }
