+++ MMOCoreORB/src/server/zone/objects/mission/CraftingMissionObjective.cpp
@@ -156,7 +156,7 @@ void CraftingMissionObjectiveImplementation::readObject(ObjectInputStream* strea
 		String _name;
 		_name.parseFromBinaryStream(stream);
 
-		uint16 _varSize = stream->readShort();
+		uint32 _varSize = stream->readInt();
 
 		int _currentOffset = stream->getOffset();
 
@@ -189,7 +189,7 @@ int CraftingMissionObjectiveImplementation::writeObjectMembers(ObjectOutputStrea
 
 	String _name;
 	int _offset;
-	uint16 _totalSize;
+	uint32 _totalSize;
 
 	return _count + 0;
 }
